src/requirements.txt
mido==1.2.9
python-rtmidi==1.4.7
RPi.GPIO==0.7.0
adafruit-circuitpython-neopixel==6.2.3
